build(nameservice-sync): target-gate the native machinery; crate builds for wasm32
The proxy clients, HttpTransport seam, and CID integrity verification
are runtime-agnostic; everything else — sync/pull/clone, SSE watchers,
config store, origin fetchers, ReqwestTransport — is native-only and now
sits behind cfg(not(target_arch = "wasm32")) with reqwest/tokio/rand/
parking_lot/async-stream moved to native-only target dependencies.
verify_object_integrity moves to a runtime-agnostic integrity module
(origin:: and crate-root re-export paths unchanged); the default
constructors that build ReqwestTransport are native-only, wasm callers
inject a transport via from_api_base_with_transport. fluree-db-nameservice
is consumed with default-features=false behind a new default 'native'
feature, so native dependency resolution is byte-identical.

CI: the wasm32 job now checks -p fluree-db-nameservice-sync alongside
fluree-db-api, so the crate's wasm buildability can no longer drift
(it was previously unverified — the E-doc's claim was vacuous).

Consumer: fluree-db-browser implements HttpTransport over fetch and
builds ProxyStorage/ProxyNameService on wasm32 via the injection
constructors.
