+++ .github/workflows/ci.yml
@@ -333,20 +333,17 @@ jobs:
           tool: wasm-pack
 
       # clippy, not check: `cargo check` cannot see clippy-only lints, and
-      # nothing else runs clippy against the wasm32 target. Two of those lints
-      # (a dropped un-awaited future in the CAS bridge stub, a needless return
-      # in a wasm-cfg storage arm) sat red under this exact command while the
+      # nothing else runs clippy against the wasm32 target. Three of those
+      # lints (a dropped un-awaited future in the CAS bridge stub, a needless
+      # return in a wasm-cfg storage arm, an explicit into_iter loop in the
+      # sequential multi-query twin) sat red under this command while the
       # check-only gate reported green. clippy lints every workspace member in
-      # the -p closure, so this covers the engine crates behind fluree-db-api,
-      # not just the umbrella.
+      # the -p closure at lib level, so this covers the engine crates behind
+      # fluree-db-api (and nameservice-sync), not just the umbrellas.
       # No --all-targets: it drags in fluree-db-api's dev-dependencies, whose
-      # tokio `net`/mio does not build for wasm32 (unrelated to our code). The
-      # lib-level pass is what matters here anyway — clippy lints every
-      # workspace member in the -p closure at lib level, which is how the two
-      # latent lints this gate exists for (in fluree-db-binary-index and
-      # fluree-db-connection) surfaced.
+      # tokio `net`/mio does not build for wasm32 (unrelated to our code).
       - name: Lint engine stack (wasm32 clippy)
-        run: cargo clippy --target wasm32-unknown-unknown -p fluree-db-api --no-default-features
+        run: cargo clippy --target wasm32-unknown-unknown -p fluree-db-api -p fluree-db-nameservice-sync --no-default-features
 
       - name: Lint parse layer, no default features (wasm32 clippy)
         run: cargo clippy --target wasm32-unknown-unknown --no-default-features -p fluree-vocab -p fluree-graph-ir -p fluree-graph-turtle -p fluree-graph-json-ld -p fluree-db-sparql
