+++ fluree-db-nameservice-sync/Cargo.toml
@@ -6,33 +6,43 @@ edition.workspace = true
 license.workspace = true
 repository.workspace = true
 
+# Target split: the proxy clients (`ProxyStorage`, `ProxyNameService`), the
+# `HttpTransport` seam, and CID integrity verification are runtime-agnostic
+# and build for wasm32; the sync/pull/clone machinery, SSE watchers, and the
+# reqwest-backed transport are native-only. Native dependency resolution is
+# unchanged by the split.
 [dependencies]
 fluree-db-core = { path = "../fluree-db-core" }
-fluree-db-nameservice = { path = "../fluree-db-nameservice" }
+fluree-db-nameservice = { path = "../fluree-db-nameservice", default-features = false }
 fluree-db-novelty = { path = "../fluree-db-novelty" }
 fluree-sse = { path = "../fluree-sse" }
 
 async-trait = { workspace = true }
 futures = { workspace = true }
 http = { workspace = true }
-reqwest = { workspace = true, features = ["stream"] }
 serde = { workspace = true }
 serde_json = { workspace = true }
-tokio = { workspace = true, features = ["time", "sync"] }
 tracing = { workspace = true }
 thiserror = { workspace = true }
-rand = { workspace = true }
 bytes = { workspace = true }
-parking_lot = "0.12"
 urlencoding = { workspace = true }
-async-stream = "0.3"
 
 # Vended S3 credentials consumer (feature "aws")
 fluree-db-storage-aws = { path = "../fluree-db-storage-aws", optional = true, default-features = false, features = ["s3"] }
 aws-config = { workspace = true, optional = true }
 aws-credential-types = { workspace = true, optional = true }
 
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+reqwest = { workspace = true, features = ["stream"] }
+tokio = { workspace = true, features = ["time", "sync"] }
+rand = { workspace = true }
+parking_lot = "0.12"
+async-stream = "0.3"
+
 [features]
+default = ["native"]
+# Filesystem-backed nameservice + core (tokio). Off for wasm32 consumers.
+native = ["fluree-db-nameservice/native", "fluree-db-core/native"]
 # Consume vended S3 credentials: build an S3 reader from a grant minted by
 # a remote server's GET /storage/credentials endpoint.
 aws = ["dep:fluree-db-storage-aws", "dep:aws-config", "dep:aws-credential-types"]
