+++ fluree-db-nameservice-sync/src/origin.rs
@@ -9,35 +9,15 @@
 
 use crate::error::{Result, SyncError};
 use fluree_db_core::pack::PackRequest;
-use fluree_db_core::{ContentId, CODEC_FLUREE_COMMIT};
+use fluree_db_core::ContentId;
 use fluree_db_nameservice::NsRecord;
 use fluree_db_nameservice::{LedgerConfig, Origin};
 use tracing::{debug, warn};
 
-// ---------------------------------------------------------------------------
-// Integrity verification
-// ---------------------------------------------------------------------------
-
-/// Verify fetched object bytes against a CID, with format-sniffing for commits.
-///
-/// - Commit blobs (`FCV2` magic): SHA-256 of full blob via `verify_commit_blob`
-/// - All other kinds (txn, config, dict, index, etc.): full-bytes SHA-256
-///
-/// **Forward-compat note:** If a future commit format uses `CODEC_FLUREE_COMMIT`
-/// but has different hashing rules, add its magic-byte check here — the
-/// `id.verify(bytes)` fallback assumes full-bytes SHA-256.
-pub fn verify_object_integrity(id: &ContentId, bytes: &[u8]) -> bool {
-    const COMMIT_V2_MAGIC: &[u8] = b"FCV2";
-
-    if id.codec() == CODEC_FLUREE_COMMIT && bytes.starts_with(COMMIT_V2_MAGIC) {
-        match fluree_db_core::commit::codec::verify_commit_blob(bytes) {
-            Ok(derived_id) => derived_id == *id,
-            Err(_) => false,
-        }
-    } else {
-        id.verify(bytes)
-    }
-}
+// Integrity verification lives in the runtime-agnostic `integrity` module
+// (it is shared with the wasm32 build); re-exported here so existing
+// `origin::verify_object_integrity` paths keep working.
+pub use crate::integrity::verify_object_integrity;
 
 /// Check whether an origin's auth requirement is satisfiable with the
 /// available credentials.
